+++ stl/lib/src/primitives/int8.dart
@@ -6,9 +6,10 @@ import 'dart:typed_data';
 /// This inherently bounds the variable within standard memory specifications and rigorously
 /// guarantees that mathematical operators intuitively overflow using C++ style
 /// constraints, exactly mimicking real hardware boundaries, providing complete safety!
-extension type Int8
-/// Instantiates a dynamically allocated [Int8] bounds-checked element.
-(Int8List _data) {
+extension type Int8._(Int8List _data)  {
+  /// Instantiates a dynamically allocated [Int8] bounds-checked element.
+  Int8(this._data);
+
   /// Dynamically instantiates a [Int8] value mapped sequentially into memory.
   Int8.from(int value) : _data = Int8List(1)..[0] = value;