+++ stl/lib/stl.dart
@@ -60,6 +60,7 @@ export 'src/math/numeric.dart';
 export 'src/math/complex.dart';
 export 'src/math/number_theory.dart';
 export 'src/math/cmath.dart';
+export 'src/math/random.dart';
 export 'src/geometry/point.dart';
 export 'src/geometry/shape.dart';
 export 'src/geometry/circle.dart';